initial commit of moon-poster
this makes a neat poster showing the phases of the moon, laid out in a
logarithmic spiral for each month.  i make full moons and new moons,
along with friday and saturday nights for planning evening hikes.  this
should automatically make a poster for any year you like.

making the poster is a two-step process: first, use the python program
moon.py, which uses the pyephem package to calculate moon phases.  next,
run the plot script moon.rb, which is made for the tioga plotting
package.

if you want to make a whole bunch of posters all at once, the script
make-posters.rb will automatically make them over a specified range of
years.